// src/models/rep.ts
export interface Rep {
  name: string;
  email: string;
  role: string;
  taxid: string;
  typeofbusiness: string;
  weekstart: string;
  opentime: string;
  experience: string;
  closedtime: string;
  companylogo: string;
  companyname: string,
  contactno: number;
  companywebsite: string;
  phone: string;
  pic: string;
  weekend: string;
  address: string;
  fax: string;
  date:Date;
  active_plan:string;
  buyer:string[];
  street:string;
  city:string;
  password:string;
  rep:string[];
}